refactor(news): extract ArticlePersistenceService with REQUIRES_NEW for per-article transaction isolation (#19)
